Create main.py
I created a loop starting with "quit" to determine the operation of the game. Next, I created a number range for the player to guess, sorting different outcomes and giving hints so the player can eventually get the correct answer. After the completion, provide the players with some basic information accumulated in the game and asks the player whether wants to try again or not, thus restart or end the whole game loop.